% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nml.r
\name{nml_load_default}
\alias{nml_load_default}
\title{Load the namelist parameters needed by S3COM}
\usage{
nml_load_default()
}
\value{
list, Namelist parameters S3COM
}
\description{
Load the namelist parameters needed by S3COM
}
