% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nml.r
\name{nml_write}
\alias{nml_write}
\title{Write the S3COM namelist file}
\usage{
nml_write(fn.nml, l)
}
\arguments{
\item{fn.nml}{character, name of the output namelist file}

\item{l}{list, namelist variables created by the \code{nml_load_default} function}
}
\value{
Namelist file for S3COM binary
}
\description{
Write the S3COM namelist file
}
