% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rs3com.r
\docType{package}
\name{Rs3com-package}
\alias{Rs3com}
\alias{Rs3com-package}
\title{Welcome to the Rs3com package!}
\description{
Here we just load dplyr so that the \%>\% operator is recognized throughout
}
\author{
\strong{Maintainer}: Odran Sourdeval \email{odran.sourdeval@univ-lille.fr} (\href{https://orcid.org/0000-0002-2822-5303}{ORCID})

}
