% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3com_run.r
\name{s3com_run}
\alias{s3com_run}
\title{Run S3COM using a specified namelist file}
\usage{
s3com_run(fn_nml, path_s3com)
}
\arguments{
\item{fn_nml}{character, name of the namelist file}

\item{path_s3com}{path to the S3COM binary}
}
\description{
Run S3COM using a specified namelist file
}
